Move out IR and VM phases.
